// src/lib/supabase/database.types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
  public: {
    Tables: {
// ...
      domains: {
        Row: {
          created_at: string | null
          domain: string
          hosted: boolean | null
          metadata: Json | null
          notification_frequencies:
            | Database["public"]["Enums"]["notification_frequency"][]
            | null
          notification_threshold: number | null
          user_id: string | null
          vercel_project_id: string | null
          verification_details: Json | null
          verified: boolean | null
        }
        Insert: {
          created_at?: string | null
          domain: string
          hosted?: boolean | null
          metadata?: Json | null
          notification_frequencies?:
            | Database["public"]["Enums"]["notification_frequency"][]
            | null
          notification_threshold?: number | null
          user_id?: string | null
          vercel_project_id?: string | null
          verification_details?: Json | null
          verified?: boolean | null
        }
        Update: {
          created_at?: string | null
          domain?: string
          hosted?: boolean | null
          metadata?: Json | null
          notification_frequencies?:
            | Database["public"]["Enums"]["notification_frequency"][]
            | null
          notification_threshold?: number | null
          user_id?: string | null
          vercel_project_id?: string | null
          verification_details?: Json | null
          verified?: boolean | null
        }
        Relationships: []
      }
// ...
    }
// ...
  }
}

type PublicSchema = Database[Extract<keyof Database, "public">]

export type Tables<
  PublicTableNameOrOptions extends
    | keyof (PublicSchema["Tables"] & PublicSchema["Views"])
    | { schema: keyof Database },
  TableName extends PublicTableNameOrOptions extends { schema: keyof Database }
    ? keyof (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
        Database[PublicTableNameOrOptions["schema"]]["Views"])
    : never = never,
> = PublicTableNameOrOptions extends { schema: keyof Database }
  ? (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
      Database[PublicTableNameOrOptions["schema"]]["Views"])[TableName] extends {
      Row: infer R
    }
    ? R
    : never
  : PublicTableNameOrOptions extends keyof (PublicSchema["Tables"] &
        PublicSchema["Views"])
    ? (PublicSchema["Tables"] &
        PublicSchema["Views"])[PublicTableNameOrOptions] extends {
        Row: infer R
      }
      ? R
      : never
    : never
// ...
export type Enums<
  PublicEnumNameOrOptions extends
    | keyof PublicSchema["Enums"]
    | { schema: keyof Database },
  EnumName extends PublicEnumNameOrOptions extends { schema: keyof Database }
    ? keyof Database[PublicEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = PublicEnumNameOrOptions extends { schema: keyof Database }
  ? Database[PublicEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : PublicEnumNameOrOptions extends keyof PublicSchema["Enums"]
    ? PublicSchema["Enums"][PublicEnumNameOrOptions]
    : never